+++ cmd/main.go
@@ -8,8 +8,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/alecthomas/kong"
 	log "github.com/sirupsen/logrus"
-	flag "github.com/spf13/pflag"
 )
 
 var Version = "unknown"
@@ -18,52 +18,42 @@ var Tag = "NO-TAG"
 var CommitID = "unknown"
 var Delta = ""
 
+type CLI struct {
+	Interface      []string `kong:"short='i',help='Two or more interfaces to use'"`
+	FixedIp        []string `kong:"short='I',help='IPs to always send to iface@ip'"`
+	Port           []int32  `kong:"short='p',help='One or more UDP ports to process'"`
+	Timeout        int64    `kong:"short='t',default=250,help='Timeout in msec'"`
+	CacheTTL       int64    `kong:"short='T',default=180,help='Client IP cache TTL in minutes'"`
+	Level          string   `kong:"short='L',default='info',enum='trace,debug,info,warn,error',help='Log level [trace|debug|info|warn|error]'"`
+	LogLines       bool     `kong:"help='Print line number in logs'"`
+	Logfile        string   `kong:"default='stderr',help='Write logs to filename'"`
+	Pcap           bool     `kong:"short='P',help='Generate pcap files for debugging'"`
+	PcapPath       string   `kong:"short='d',default='/root',help='Directory to write debug pcap files'"`
+	ListInterfaces bool     `kong:"short='l',help='List available interfaces and exit'"`
+	Version        bool     `kong:"short='v',help='Print version information'"`
+}
+
+func init() {
+	log.SetFormatter(&log.TextFormatter{
+		DisableLevelTruncation: true,
+		PadLevelText:           true,
+		DisableTimestamp:       true,
+	})
+	log.SetOutput(os.Stderr)
+}
+
 func main() {
-	var _fixed_ip = []string{}
-	var interfaces = []string{}
-	var ports = []int32{}
-	var timeout int64
-	var cachettl int64
-	var debug bool
-	var logfile string
-	var version bool
-	var ilist bool
-	var fixed_ip = map[string][]string{}
+	cli := CLI{}
+	parser := kong.Must(
+		&cli,
+		kong.Name("udp-proxy-2020"),
+		kong.Description("A crappy UDP proxy for the year 2020 and beyond!"),
+		kong.UsageOnError(),
+	)
+	_, err := parser.Parse(os.Args[1:])
+	parser.FatalIfErrorf(err)
 
-	// option parsing
-	flag.StringSliceVar(&interfaces, "interface", []string{}, "Two or more interfaces to use")
-	flag.StringSliceVar(&_fixed_ip, "fixed-ip", []string{}, "IPs to always send to: iface@ip")
-	flag.Int32SliceVar(&ports, "port", []int32{}, "One or more UDP ports to process")
-	flag.Int64Var(&timeout, "timeout", 250, "Timeout in msec")
-	flag.Int64Var(&cachettl, "cachettl", 3*60, "Client IP cache TTL in min")
-	flag.BoolVar(&debug, "debug", false, "Enable debugging")
-	flag.StringVar(&logfile, "logfile", "stderr", "Log to file instead of stderr")
-	flag.BoolVar(&ilist, "list-interfaces", false, "List available interfaces and exit")
-	flag.BoolVar(&version, "version", false, "Print version and exit")
-
-	flag.Parse()
-
-	// log.DisableLevelTruncation(true) <-- supposed to work, but doesn't?
-
-	// turn on debugging?
-	if debug == true {
-		log.SetReportCaller(true)
-		log.SetLevel(log.DebugLevel)
-	} else {
-		log.SetLevel(log.WarnLevel)
-	}
-
-	if logfile == "stderr" {
-		log.SetOutput(os.Stderr)
-	} else {
-		file, err := os.OpenFile(logfile, os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0644)
-		if err != nil {
-			log.WithError(err).Fatalf("Unable to open log file: %s", logfile)
-		}
-		log.SetOutput(file)
-	}
-
-	if version == true {
+	if cli.Version {
 		delta := ""
 		if len(Delta) > 0 {
 			delta = fmt.Sprintf(" [%s delta]", Delta)
@@ -74,20 +64,49 @@ func main() {
 		os.Exit(0)
 	}
 
-	if ilist == true {
+	// Setup Logging
+	switch cli.Level {
+	case "trace":
+		log.SetLevel(log.TraceLevel)
+	case "debug":
+		log.SetLevel(log.DebugLevel)
+	case "warn":
+		log.SetLevel(log.WarnLevel)
+	case "info":
+		log.SetLevel(log.InfoLevel)
+	case "error":
+		log.SetLevel(log.ErrorLevel)
+	}
+
+	if cli.LogLines {
+		log.SetReportCaller(true)
+	}
+
+	if cli.ListInterfaces {
 		listInterfaces()
 		os.Exit(0)
 	}
 
-	// Neeed at least two interfaces
-	if len(interfaces) < 2 {
-		log.Fatal("Please specify two or more interfaces via --interface")
+	if len(cli.Interface) < 2 {
+		log.Fatalf("Please specify two or more --interface")
+	}
+	if len(cli.Port) < 1 {
+		log.Fatalf("Please specify one or more --port")
+	}
+
+	if cli.Logfile != "stderr" {
+		file, err := os.OpenFile(cli.Logfile, os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0644)
+		if err != nil {
+			log.WithError(err).Fatalf("Unable to open log file: %s", cli.Logfile)
+		}
+		log.SetOutput(file)
 	}
 
 	// handle our timeout
-	to := parseTimeout(timeout)
+	to := parseTimeout(cli.Timeout)
 
-	for _, fip := range _fixed_ip {
+	var fixed_ip = map[string][]string{}
+	for _, fip := range cli.FixedIp {
 		split := strings.Split(fip, "@")
 		if len(split) != 2 {
 			log.Fatalf("--fixed-ip %s is not in the correct format of <interface>@<ip>", fip)
@@ -95,7 +114,7 @@ func main() {
 		if net.ParseIP(split[1]) == nil {
 			log.Fatalf("--fixed-ip %s IP address is not a valid IPv4 address", fip)
 		}
-		if !stringInSlice(split[0], interfaces) {
+		if !stringInSlice(split[0], cli.Interface) {
 			log.Fatalf("--fixed-ip %s interface must be specified via --interface", fip)
 		}
 		fixed_ip[split[0]] = append(fixed_ip[split[0]], split[1])
@@ -104,7 +123,7 @@ func main() {
 	// create our Listeners
 	var seenInterfaces = []string{}
 	var listeners = []Listen{}
-	for _, iface := range interfaces {
+	for _, iface := range cli.Interface {
 		// check for duplicates
 		if stringPrefixInSlice(iface, seenInterfaces) {
 			log.Fatalf("Can't specify the same interface (%s) multiple times", iface)
@@ -117,13 +136,25 @@ func main() {
 		}
 
 		var promisc bool = (netif.Flags & net.FlagBroadcast) == 0
-		listeners = append(listeners, newListener(netif, promisc, ports, to, fixed_ip[iface]))
+		l := newListener(netif, promisc, cli.Port, to, fixed_ip[iface])
+		listeners = append(listeners, l)
 	}
 
 	// init each listener
-	ttl, _ := time.ParseDuration(fmt.Sprintf("%dm", cachettl))
+	ttl, _ := time.ParseDuration(fmt.Sprintf("%dm", cli.CacheTTL))
 	for i := range listeners {
 		initializeInterface(&listeners[i])
+		if cli.Pcap {
+			if fName, err := listeners[i].OpenWriter(cli.PcapPath, In); err != nil {
+				log.Fatalf("Unable to open pcap file %s: %s", fName, err.Error())
+			}
+			if fName, err := listeners[i].OpenWriter(cli.PcapPath, Out); err != nil {
+				log.Fatalf("Unable to open pcap file %s: %s", fName, err.Error())
+			}
+			if fName, err := listeners[i].OpenWriter(cli.PcapPath, InOut); err != nil {
+				log.Fatalf("Unable to open pcap file %s: %s", fName, err.Error())
+			}
+		}
 		listeners[i].clientTTL = ttl
 		defer listeners[i].handle.Close()
 	}